build - Use of npm as package manager to avoid some problems with retest
There was an issue in the testing library of retest that it was not
able find the necessary modules in the node_modules folder with pnpm.
In addition with deno the test was not able to pass in the part of the
evaluation of the event listeners. I was not able to find the error.
My speculation is that jsdom did not recognize the window global object
that only exists in the browser.